moved FilingTypes to common constants
