+++ src/enums/index.ts
@@ -9,7 +9,6 @@ export * from './entity-states'
 export * from './fee-summary-actions'
 export * from './filing-codes'
 export * from './filing-names'
-export * from './filing-types'
 export * from './jurisdiction-location'
 export * from './name-request-codes'
 export * from './name-request-states'
