+++ src/enums/sbc-common-components-constants.ts
@@ -160,6 +160,42 @@ export enum AffiliationTypes {
   REGISTRATION = 'Registration'
 }
 
+export enum FilingTypes {
+  ADMIN_FREEZE = 'adminFreeze',
+  ALTERATION = 'alteration',
+  AGM_EXTENSION = 'agmExtension',
+  AGM_LOCATION_CHANGE = 'agmLocationChange',
+  AMALGAMATION_APPLICATION = 'amalgamationApplication',
+  AMALGAMATION_OUT = 'amalgamationOut',
+  ANNUAL_REPORT = 'annualReport',
+  CHANGE_OF_ADDRESS = 'changeOfAddress',
+  CHANGE_OF_COMPANY_INFO = 'changeOfCompanyInfo', // a type of alteration
+  CHANGE_OF_DIRECTORS = 'changeOfDirectors',
+  CHANGE_OF_NAME = 'changeOfName',
+  CHANGE_OF_REGISTRATION = 'changeOfRegistration',
+  CONSENT_AMALGAMATION_OUT = 'consentAmalgamationOut',
+  CONSENT_CONTINUATION_OUT = 'consentContinuationOut',
+  CONTINUATION_IN = 'continuationIn',
+  CONTINUATION_OUT = 'continuationOut',
+  CONVERSION = 'conversion',
+  CORRECTION = 'correction',
+  COURT_ORDER = 'courtOrder',
+  DISSOLUTION = 'dissolution',
+  DISSOLVED = 'dissolved',
+  INCORPORATION_APPLICATION = 'incorporationApplication',
+  NOTICE_OF_WITHDRAWAL = 'noticeOfWithdrawal',
+  PUT_BACK_ON = 'putBackOn',
+  PUT_BACK_OFF = 'putBackOff',
+  REGISTRATION = 'registration',
+  REGISTRARS_NOTATION = 'registrarsNotation',
+  REGISTRARS_ORDER = 'registrarsOrder',
+  RESTORATION = 'restoration',
+  SPECIAL_RESOLUTION = 'specialResolution',
+  TRANSITION = 'transition',
+
+  UNKNOWN = 'unknown' // for fallback
+}
+
 export enum LearFilingTypes {
   AMALGAMATION = 'Amalgamation',
   INCORPORATION = 'Incorporation',
